tests(app): Add failing test for cube search navigation bug
