+++ packages/openneuro-app/src/scripts/search/__tests__/search-params-ctx.spec.tsx
@@ -2,20 +2,9 @@ import React from 'react'
 import { render, screen, fireEvent } from '@testing-library/react'
 import { SearchParamsProvider, SearchParamsCtx } from '../search-params-ctx'
 import AuthorInput from '../inputs/author-input'
+import initialSearchParams from '../initial-search-params'
 import { MemoryRouter } from 'react-router-dom'
 import '@testing-library/jest-dom'
-
-/**
- * Render with SearchParamsCtx state
- */
-export const searchRender = (ui, { providerProps, ...renderOptions }) => {
-  return render(
-    <SearchParamsCtx.Provider {...providerProps}>
-      {ui}
-    </SearchParamsCtx.Provider>,
-    renderOptions,
-  )
-}
 
 describe('SearchParamsProvider', () => {
   it('restores URL searchParams state', () => {
@@ -60,5 +49,30 @@ describe('SearchParamsProvider', () => {
     expect(screen.getByText(/^Received:/).textContent).toBe(
       'Received: Test Author, New Author',
     )
+  })
+  it('setSearchParams is callable with object argument', () => {
+    const route = '/search?query={"authors"%3A["Test+Author"]}'
+    const wrapper = ({ children }) => (
+      <MemoryRouter initialEntries={[route]}>
+        <SearchParamsProvider>{children}</SearchParamsProvider>
+      </MemoryRouter>
+    )
+
+    let setSearchParams
+
+    render(
+      <SearchParamsCtx.Consumer>
+        {value => {
+          setSearchParams = value.setSearchParams
+          return <span>Received: {value.searchParams.authors.pop()}</span>
+        }}
+      </SearchParamsCtx.Consumer>,
+      { wrapper },
+    )
+    expect(screen.getByText(/^Received:/).textContent).toBe(
+      'Received: Test Author',
+    )
+
+    expect(() => setSearchParams(initialSearchParams)).not.toThrow()
   })
 })
